requirements.txt
streamlit
pandas
numpy
scikit-learn
python-binance
joblib
altair

